pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "norag"
version = "0.1.0"
description = "Knowledge compiler. Compile documents into machine-optimized knowledge. No vectors. No chunks. No embeddings."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
keywords = ["knowledge", "llm", "compiler", "rag", "cli"]

dependencies = [
  "pydantic>=2.7",
  "typer>=0.12",
  "pyyaml>=6.0",
  "httpx>=0.27",
  "rich>=13.0",
  "pymupdf>=1.24",
]

[project.optional-dependencies]
claude = ["anthropic>=0.40"]
ollama = ["ollama>=0.3"]
dev = ["pytest>=8.0", "ruff>=0.4", "mypy>=1.10"]

[project.scripts]
norag = "norag.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/norag"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
